# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Geoprofs API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/verlof/beoordeling": {
            "get": {
                "tags": [
                    "Verlof"
                ],
                "summary": "Alle verlofaanvragen ophalen voor beoordeling",
                "operationId": "3f6bfe5977392d35adc626ebe37d3a5b",
                "responses": {
                    "200": {
                        "description": "Lijst van aanvragen"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/verlof/beoordeling/{aanvraag}/accept": {
            "post": {
                "tags": [
                    "Verlof"
                ],
                "summary": "Accepteer een verlofaanvraag",
                "operationId": "f8ba78118190164e2b5e075c004b0e75",
                "parameters": [
                    {
                        "name": "aanvraag",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesvol geaccepteerd"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/verlof/beoordeling/{aanvraag}/reject": {
            "post": {
                "tags": [
                    "Verlof"
                ],
                "summary": "Verlofaanvraag afwijzen",
                "operationId": "390513ce98fc74003604f91a2bc4cfc5",
                "parameters": [
                    {
                        "name": "aanvraag",
                        "in": "path",
                        "description": "ID van de verlofaanvraag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verlofaanvraag afgewezen"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mijn-aanvragen": {
            "get": {
                "tags": [
                    "Verlof"
                ],
                "summary": "Haal de verlofaanvragen van de ingelogde medewerker op",
                "operationId": "a1824db35e7e305c7ceefc9055dcb7a2",
                "responses": {
                    "200": {
                        "description": "Lijst van verlofaanvragen succesvol opgehaald",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "medewerker_id": {
                                                "type": "integer",
                                                "example": 42
                                            },
                                            "type": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "naam": {
                                                        "type": "string",
                                                        "example": "Ziekteverlof"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "aanvraag_datum": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-11-04"
                                            },
                                            "start_datum": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-11-10"
                                            },
                                            "eind_datum": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-11-12"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "In afwachting"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Niet geautoriseerd"
                    },
                    "500": {
                        "description": "Serverfout"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token voor authenticatie",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}